+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -1,9 +1,9 @@
 object Dependencies {
-    const val material = "com.google.android.material:material:1.4.0-beta01"
+    const val material = "com.google.android.material:material:1.4.0-rc01"
     const val desugar = "com.android.tools:desugar_jdk_libs:1.1.5"
 
     object ComposeMaterialDialogs {
-        const val version = "0.4.1"
+        const val version = "0.4.3"
 
         const val core = "io.github.vanpra.compose-material-dialogs:core:$version"
         const val datetime = "io.github.vanpra.compose-material-dialogs:datetime:$version"
@@ -15,7 +15,7 @@ object Dependencies {
     }
 
     object Accompanist {
-        private const val version = "0.11.1"
+        private const val version = "0.12.0"
         const val pager = "com.google.accompanist:accompanist-pager:$version"
     }
 
@@ -32,7 +32,7 @@ object Dependencies {
 
     object AndroidX {
         const val appcompat = "androidx.appcompat:appcompat:1.4.0-alpha02"
-        const val coreKtx = "androidx.core:core-ktx:1.6.0-beta02"
+        const val coreKtx = "androidx.core:core-ktx:1.6.0-rc01"
 
         object Testing {
             const val version = "1.4.0-alpha06"
@@ -42,16 +42,16 @@ object Dependencies {
         }
 
         object Compose {
-            const val version = "1.0.0-beta08"
+            const val version = "1.0.0-beta09"
 
             const val ui = "androidx.compose.ui:ui:$version"
             const val material = "androidx.compose.material:material:$version"
             const val materialIconsExtended =
                 "androidx.compose.material:material-icons-extended:$version"
-            const val testing = "androidx.compose.ui:ui-test-junit4:$version"
 
-            const val activity = "androidx.activity:activity-compose:1.3.0-beta01"
-            const val navigation = "androidx.navigation:navigation-compose:2.4.0-alpha02"
+            const val testing = "androidx.compose.ui:ui-test-junit4:$version"
+            const val activity = "androidx.activity:activity-compose:1.3.0-beta02"
+            const val navigation = "androidx.navigation:navigation-compose:2.4.0-alpha03"
         }
     }
 }